Fix Python 3 support

+++ colorful/utils.py
@@ -19,4 +19,4 @@ def hex_to_rgb(value):
     value = value.lstrip('#')
 
     length = len(value)
-    return tuple(int(value[i:i + length / 3], 16) for i in range(0, length, length / 3))
+    return tuple(int(value[i:i + int(length / 3)], 16) for i in range(0, length, int(length / 3)))
